scanner/requirements.txt
opencv-python
pyzbar
pyserial
requests
